Remove lazy_static dependency
The main use of lazy_static for microhop is to have global Strings to
represent mountpoints, but this can bee solved by using &'static str.
Some functions however dynamically construct different mountpoints, so
make mount_fs() generic, so it accepts both &str and String.

In order to further simplify things, introduce a new struct to represent
a system directory to be mounted instead of using 3-tuples.